chore(settings): 전역 저장소와 타입 구조 정리
